% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\title{Bathymetric raster}
\format{A \code{RasterLater} object. It has a resolution of 10x10 m and a
    total extension of 221x221 cells. Land areas are denoted by \code{NA}
    values.}
\usage{
bathymetry
}
\description{
Bathymetric raster of the Medes Islands marine reserve (Catalonia, NW
Mediterranean Sea). The coordinates are in UTM, referred to the datum WGS84
zone 31N.
}
\keyword{datasets}
