% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatChronPlot.R
\name{colorScale}
\alias{colorScale}
\title{Plot the color scale in a spatial chronogram plot}
\usage{
colorScale(labels, colors, title, title.adj = -2.8, size = 1,
  cex.lab = 0.9)
}
\arguments{
\item{labels}{labels (receiver IDs) for each color cathegory.}

\item{colors}{vector with the color for each receiver.}

\item{title}{title for the scale bar.}

\item{title.adj}{ajustment of the title of the scale bar.}

\item{size}{Height of the scale from the bottom to the top, as percentage of
the total height.}

\item{cex.lab}{the magnification to be used for the labels.}
}
\value{
Plots the scale bar in the right of a spatial chronogram plot.
}
\description{
Plots a color scale bar in the right of the plot. It usually requires to
increase the right margin to fit.
}
