% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatChronPlot.R
\name{dateAxis}
\alias{dateAxis}
\title{Plot the X axis of a spatial chronogram plot (dates)}
\usage{
dateAxis(xlim, spacing = c("auto", "month", "day"))
}
\arguments{
\item{xlim}{the minimum and maximum values for the axis in \code{Date}
format.}

\item{spacing}{tick spacing. If \code{NULL}, it will be automatically
choosen depending on the length of the period to plot. It can be set to
'day' or 'month'.}
}
\value{
Plots the X axis in a spatial chronogram plot.
}
\description{
Plots the date in the X axis of an spatial chronogram plot.
}
