% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depth_cost_list}
\alias{depth_cost_list}
\title{List of transition layers}
\format{A \code{list} of \code{TransitionLayer}, with one element for each
    1 m depth-interval between 0 and 70 m, for the Medes Islands MPA. Each
    \code{TransitionLayer} excludes the cells at depths shallower than the
    depth-interval that defines each element in the list.}
\usage{
depth_cost_list
}
\description{
Explanation of the depth cost list
}
\keyword{datasets}
