% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatChronPlot.R
\name{detect2matrix}
\alias{detect2matrix}
\title{Convert detection data into a matrix}
\usage{
detect2matrix(time.stamp, rec.id, time.int = "30min", xlim = NULL)
}
\arguments{
\item{time.stamp}{vector with the date and time of each detection in
\code{POSIXt} format.}

\item{rec.id}{vector with the ID codes of the receivers for each detection.}

\item{time.int}{time interval for the Y axis of the plot, as given to the
\code{units} argument of the \code{\link[lubridate]{round_date}}
function in the \code{lubridate} package.}

\item{xlim}{date limits for the x axis in \code{Date} format.}
}
\value{
A matrix with the IDs of the receivers for each day and time
    interval.
}
\description{
This function takes the original detections from acoustic telemetry data and
chooses the receiver with the largest amount of detections for each day and
a given time interval (e.g. 30 min).
}
