% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelize.R
\name{interpPath}
\alias{interpPath}
\title{Interpolate trajectory into points}
\usage{
interpPath(path, time.step = 2, max.lag = NULL)
}
\arguments{
\item{path}{a data.frame with the trajectory of a individual. It must
contain the \code{time.stamp}, \code{x}, \code{y} and \code{depth}
columns.}

\item{time.step}{time interval into which interpolate the trajectory, in
minutes.}

\item{max.lag}{maximum time period, in hours, above which the trajectory
between two locations will not be interpolated.}
}
\value{
a \code{data.frame} with the resulting interpolated locations.
}
\description{
This function takes a single trajectory and interpolates it into a set of
points that are spaced by a given time step. If a maximum time lag is
provided, the function will not interpolate segments between locations that
are separated by more than the maximum lag.
}
