% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntPath.R
\name{leastCostMap}
\alias{leastCostMap}
\title{Generate a least cost transition matrix from a topographical raster}
\usage{
leastCostMap(topo, min.depth = 0, max.depth = NULL)
}
\arguments{
\item{topo}{raster dataset (\code{RasterLayer} object) with the topographic
information of the study area (bathymetry or elevation).}

\item{min.depth, max.depth}{Minimum and maximum depths that the path is
allowed to cross. If not provided, only areas that are marked with
\code{NA} in the topographic raster will be avoided.}
}
\value{
A \code{`TransitionLayer`} object (\code{gdistance} package).
}
\description{
This function generates a \code{`TransitionLayer`} (\code{gdistance}
package) that excludes the cells that are outside a depth range, which is
used later to calculate the shortest path between two points that avoids
the excluded cells.
}
\examples{
library(raster)
dist.cost <- leastCostMap(bathymetry, min.depth = 30)
plot(raster(dist.cost))


}
