% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap3d.R
\name{overlap3d}
\alias{overlap3d}
\title{Spatial overlap between several utilization distributions}
\usage{
overlap3d(ud.vol.list, level, symmetric = TRUE)
}
\arguments{
\item{ud.vol.list}{list of \code{RasterLayer}, \code{RasterStack} or
\code{RasterBrick} objects with the UDs volumes to overlap. If the
elements of the list are \code{RasterStack} or \code{RasterBrick}
objects, the number and name of the layers must coincide.}

\item{level}{UD volume probability contour to be used to calculate
volume overlaps.}

\item{symmetric}{logical. If \code{TRUE}, the overlap indexes are calculated
referred to the total joint volume of each pair of UD volumes. If
\code{FALSE}, two overlap indexes are calculated for each pair of UDs,
each one referred to the volume of one of the UDs to compare (see the
\link{volOverlap} function.}
}
\value{
The function returns a matrix with the overlap indexes between each
    pair of UD volumes. The matrix will be symmetric if
    \code{symmetric == TRUE} or assymetric if \code{symmetric == FALSE}.
}
\description{
This function takes a list of \code{RasterLayer}, \code{RasterStack} or
\code{RasterBrick} with utilization distribution (UDs) volumes (calculated
with \link{volumeUD}) and calculates the spatial overlap within a probability
contour between each pair of them.
}
