% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeUD.R
\name{predictKde}
\alias{predictKde}
\title{Predict utilization distribution (UD) values from a \code{kde} object}
\usage{
predictKde(kde, raster, depths)
}
\arguments{
\item{kde}{a kde object.}

\item{raster}{raster to extract the horizontal coordinates of the 3D grid.}

\item{depths}{vector with depths values defining the vertical resolution of
the 3D grid.}
}
\value{
A \code{RasterStack} object with the UD volumes of a different depth
    interval in each layer.
}
\description{
This function takes a \code{kde} object and predicts the utilization
distribution (UD) values for a 3D grid defined by a raster (horizontal
resolution) and a sequence of depths (vertical resolution.)
}
