% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntPath.R
\name{probRaster}
\alias{probRaster}
\title{Create a raster of detection probabilities around a receiver}
\usage{
probRaster(dist.rast, tag.depth, ac.range.mod, topo, dist.bottom = NULL,
  max.dist = NULL, rec.depth = 5, max.depth.diff = NULL)
}
\arguments{
\item{dist.rast}{\code{RasterLayer} object with the distances from the
location of one of the receivers to each of the raster cells. Acoustic
shadows and land areas must be identified with \code{NA} values.}

\item{tag.depth}{single depth value at which the fish was observed. The
probability of being detected in shallower cells will be set to 0.}

\item{ac.range.mod}{logistic regression model of the relationship between
the detection probability and the distance from the receiver. It must be
a \code{glm} object.}

\item{topo}{raster dataset (\code{RasterLayer} object) with the topographic
information of the study area (bathymetry or elevation).}

\item{dist.bottom}{\code{RasterLayer} object with distances from a point
(the last sampled coordinate), taking into account and avoiding emerged
land areas. It is an optional argument, but it must be provided together
with the \code{max.dist} argument.}

\item{max.dist}{maximum distance that the fish is allowed to swim from the
previous location to the current one. It is calculated depending on the
maximum speed of the fish and the time elapsed between detections. It is
an optional argument, but must be provided together with the
\code{dist.bottom} argument.}

\item{rec.depth}{depth at which the receiver is placed. It is used to
incorporate the vertical distance between the transmitter and the
receiver before applying the acoustic range model.}

\item{max.depth.diff}{optional argument to establish a maximum depth (below
the \code{depth} value), below which exclude the coordinate sampling.}
}
\value{
A \code{RasterLayer} object with the spatial probability of being
    detected by a receiver.
}
\description{
Combines a distance raster for a receiver, the depth of the detection, and
an acoustic range model to generate a new \code{Rasterlayer} with the
probabilities of being detected around a receiver. It also requires a
topographic raster to exclude shallow cells. Optionally, if a raster of
distances from a point and a maximum distance is provided, excludes the
cells beyond that distance.
}
\examples{
library(raster)

# Create the acoustic range model
data(range_test)
det.range <- glm(det.ratio ~ dist.m, data = range_test,
                 family = quasibinomial(logit))

# Probability raster for the receiver 'X1'
rec <- "X1"
prob.rast <- probRaster(dist.rast = viewshed[[rec]], tag.depth = 10,
                        topo = bathymetry, ac.range.mod = det.range)
plot(prob.rast)


}
