% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{range_test}
\alias{range_test}
\title{Data from an acoustic signal range test experiment}
\format{A data frame, where each row corresponds to a 1 h interval, and with
    the following columns:
   \describe{
       \item{\code{date.time}}{date and time of each 1h interval}
       \item{\code{tag.id}}{ID of the acoustic transmitter}
       \item{\code{rec.id}}{ID of acoustic receivers}
       \item{\code{dist.m}}{distance between the receiver and the
           transmitter, in m}
       \item{\code{rec.depth.m}}{depth at which the receiver was placed. For
           each distance from the transmitter, two receivers were placed at
           different depths (10 and 12 m)}
       \item{\code{detected}}{number of acoustic signals detected by the
           receiver in each 1h interval}
       \item{\code{emitted}}{total number of signals emitted by the
           transmitter in each  1h interval. It is calculated as the amount
           of signals detected by the receiver that was placed just next to
           the transmitter (\code{dist.m = 0})}
       \item{\code{det.ratio}}{acoustic efficiency ratio, i.e. ratio between
           the signals detected by the receivers at different distances and
           the total number of signals originally emitted by the
           transmitter}
    }}
\source{
Aspillaga, E., 2017. \emph{Movement ecology of coastal fishes in a marine
protected area: implications for management and conservation}. Doctoral
thesis. Universitat de Barcelona. 218 pp.
}
\usage{
range_test
}
\description{
Data corresponding to an acoustic signal range test experiment performed in
the Medes Islands marine reserve (Catalonia, NW Mediterranean Sea). The
experiment was performed by placing receivers at 5 different distances
from one acoustic transmitter during a 24 hour period.
}
\keyword{datasets}
