% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatChronPlot.R
\name{recColor}
\alias{recColor}
\title{Assign colors to receivers}
\usage{
recColor(unique.rec.id = NULL, palette = NULL)
}
\arguments{
\item{unique.rec.id}{vector with the unique names of the receivers.}

\item{palette}{a color palette to extract the colors. If not provided,
a default palette is generated and used.}
}
\value{
A vector of color names, where the name of each element corresponds
    to the name of a receiver.
}
\description{
This function assigns a color to each of the unique receiver values using a
provided or predefined color palette.
}
