% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{receivers}
\alias{receivers}
\title{Location of acoustic receivers in the Medes Islands}
\format{Data frame with the following columns:
   \describe{
       \item{\code{id}}{ID number of each receiver}
       \item{\code{long.utm}}{longitude of the coordinate, in m}
       \item{\code{lat.utm}}{latitude of the coordinate, in m}
    }}
\source{
Aspillaga, E., 2017. \emph{Movement ecology of coastal fishes in a marine
protected area: implications for management and conservation}. Doctoral
thesis. Universitat de Barcelona. 218 pp.
}
\usage{
receivers
}
\description{
Data frame with the geographical coordinates of the 17 acoustic receivers
placed around the Medes Islands marine reserve (Catalonia, NW Mediterranean
Sea). Coordinates are in UTM, referred to the datum WGS84 zone 31N.
}
\keyword{datasets}
