% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntPath.R
\name{sampleCoord}
\alias{sampleCoord}
\title{Sample pairs of coordinates from a probability raster}
\usage{
sampleCoord(prob.raster, n = 1)
}
\arguments{
\item{prob.raster}{\code{RasterLayer} object with the probabilities of being
detected by one receiver, as given by the \code{\link{probRaster}}
function.}

\item{n}{number of pair of coordinates to return.}
}
\value{
A data frame with the \code{x} and \code{y} coordinates of the
    sampled points, in the same geographic reference system as the
    \code{prob.rast} object.
}
\description{
This function takes a raster with the detection probabilities around a
receiver and samples \code{n} number of geografic coordinates according to
to the probabilities.
}
\examples{
library(raster)

# Create the acoustic range model
data(range_test)
det.range <- glm(det.ratio ~ dist.m, data = range_test,
                 family = quasibinomial(logit))

# Probability raster for the receiver 'X1'
rec <- "X1"
prob.rast <- probRaster(dist.rast = viewshed[[rec]], tag.depth = 10,
                        topo = bathymetry, ac.range.mod = det.range)
plot(prob.rast)

# Sample coordinates
sampled.points <- sampleCoord(prob.rast, 10)
points(sampled.points, pch = "+")


}
