% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntPath.R
\name{shortPath}
\alias{shortPath}
\title{Get the shortest path between two points}
\usage{
shortPath(from, to, depth.cost, simp.tol = 10)
}
\arguments{
\item{from, to}{vectors with the (x, y) coordinates of the initial and final
points}

\item{depth.cost}{\code{TransitionLayer} returned by \link{leastCostMap}
function.}

\item{simp.tol}{Tolerance (\code{tol}) argument for the
\link[rgeos]{gSimplify} function.}
}
\value{
A data frame with the \code{x} and \code{y} coordinates defining the
    shortest path between the two points, in the same geographic reference
    system as the \code{depth.cost} object.
}
\description{
Calculates and returns the coordinates of the shortest path between two
points, taking into account a \code{TransitionLayer} generated by the
\code{leastCostMap} function.
}
\examples{
\dontrun{
library(raster)

depth.cost <- leastCostMap(bathymetry, min.depth = 20, max.depth = 30)
plot(raster(depth.cost))

from <- c(518201, 4655442)
to <-  c(518290, 4654591)

# Points can be also selected by directly clicking in the map
# from <- locator(1)
# to <- locator(1)

path <- shortPath(from = from, to = to, depth.cost = depth.cost)

points(rbind(from, to))
lines(path)
}


}
