% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatChronPlot.R
\name{spatChronPlot}
\alias{spatChronPlot}
\title{Make spatial chronogram plots with acoustic telemetry data}
\usage{
spatChronPlot(rec.id, time.stamp, time.int = "30min", colors = NULL,
  include.all.rec = TRUE, palette = NULL, xlim = NULL, xaxs = TRUE,
  yaxs = TRUE, scale = TRUE, cex.lab = 0.9)
}
\arguments{
\item{rec.id}{vector with the ID codes of the receivers for each detection.}

\item{time.stamp}{vector with the date and time of each detection in
\code{POSIXt} format.}

\item{time.int}{Time interval for the Y axis of the plot as given to the
\code{units} argument of the \code{\link[lubridate]{round_date}}
function in the \code{lubridate} package. Default is set to
\code{'30min'}.}

\item{colors}{optional \code{character} vector with the names of the colors
to plot each receiver. The names of the elements in the vector must
match with the receiver IDs in \code{rec.id}. If this argument is not
provided, colors will be automatically assigned using a color palette,
which can be provided in the \code{palette} argument.}

\item{include.all.rec}{logical, if \code{TRUE}, all the levels of the
receiver IDs (in \code{rec.id} or \code{names(colors)}) will be included
in the assignation of colors and when plotting the scale bar. If
\code{FALSE}, only the receivers that appear in \code{rec.id} will
appear in the scale bar.}

\item{palette}{if \code{colors} is not provided, a color palette can be
specified to extract the colors. If \code{colors} or \code{palette} are
not provided, a predefined color palette will be used.}

\item{xlim}{minimum and maximum date limits for the X axis, in \code{Date}
format.}

\item{xaxs, yaxs, scale}{Logical arguments to plot (if \code{TRUE}) the X and
Y axis and the lateral scale bar, respectively.}

\item{cex.lab}{the magnification to be used for the labels of the scale bar.}
}
\value{
The funtion makes a spatial chronogram plot.
}
\description{
This function creates spatial chronogram plots to visualize passive acoustic
telemetry data. Spatial chronogram plots represent the receivers with the
largest number of receptions within predefined time intervals (e.g. 30 min)
in each day of the tracking period. They are an effective way to visualize,
on a fine temporal scale, presences and absences of an individual among
different zones of the receiver array.
}
\examples{

data(tracking)
tracking$rec.id <- factor(tracking$rec.id)
tracking <- split(tracking, tracking$tag.id)

par(mfrow = c(2, 1), mar = c(3.1, 4.1, 2.1, 4.1))
spatChronPlot(rec.id = tracking[[1]]$rec.id,
              time.stamp = tracking[[1]]$time.stamp)
title(main = names(tracking)[1])
spatChronPlot(rec.id = tracking[[2]]$rec.id,
              time.stamp = tracking[[2]]$time.stamp)
title(main = names(tracking)[1])


}
\references{
Aspillaga, E., F. Bartumeus, C. Linares, R.M. Starr, À. López-Sanz,
    D. Díaz, M. Zabala, and B. Hereu. 2016. Ordinary and extraordinary
    movement behaviour of small resident fish within a Mediterranean marine
    protected area. \emph{PLoS ONE}, 11: e0159813.
}
