% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synt_path_x100}
\alias{synt_path_x100}
\title{100 synthetic random trajectories}
\format{An object of class \code{list} of length 2.}
\usage{
synt_path_x100
}
\description{
A list containing 100 random path realizations for each of the two common
dentex individuals in the sample telemetry data (\link{tracking} object).
}
\keyword{datasets}
