% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntPath.R
\name{syntPath}
\alias{syntPath}
\title{Generate synthetic trajectories of fishes from acoustic telemetry data}
\usage{
syntPath(track.data, topo, dist.rec, ac.range.mod, max.vel = 1,
  depth.cost.list = NULL, check = FALSE)
}
\arguments{
\item{track.data}{\code{data.frame} with acoustic telemetry data containing
the following columns:
    \describe{
        \item{\code{rec.id}}{ID of the acoustic receiver that registered
            each detection}
        \item{\code{time.stamp}}{date and time of each detection, in
            \code{POSIXt} format}
        \item{\code{depth}}{depth of the fish, provided by the acoustic
            transmitters in the coded signals}
        \item{\code{tag.id}}{ID of the tagged individual (Optional)}
    }}

\item{topo}{raster dataset (\code{RasterLayer} object) with the topographic
information of the study area (bathymetry or elevation).}

\item{dist.rec}{\code{RasterStack} object with the distances between all the
cells and a given receiver in each layer. Acoustic shadows and other
areas to exclude (e.g. land areas) must be identified with \code{NA}
values.}

\item{ac.range.mod}{logistic regression model of the relationship between
the detection probability and the distance from the receiver. It must be
a \code{glm} object.}

\item{max.vel}{Maximum average velocity at which the animal is supposed to
move. This argument is used to limit the maximum space between sampled
locations. By default is set to 1 m/s.}

\item{depth.cost.list}{A list containing \code{TransitionLayer} objects,
each one excluding the cells placed in areas shallower thant the depth
that gives the name to the element in the list.}

\item{check}{Print a message after sampling each coordinate.}
}
\value{
A \code{data.frame} containing the time stamp, receiver, coordinates
    and depth of each sampled location.
}
\description{
This function generates a stochastic trajectory of a fish by sampling a
possible pair of coordinates for each time step. It takes into account the
depth of the fish, the probability of being detected as function of the
distance to the receiver (including acoustic shadow areas outside the
acoustic range), and the topography of the study area.
}
