% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tCol.R
\name{tCol}
\alias{tCol}
\title{Create colors with transparency}
\usage{
tCol(color, trans = 50, name = NULL)
}
\arguments{
\item{color}{character vector with the names or codes of the original
colors.}

\item{trans}{percentage of transparency to apply to the colors (0 to 100).
If a unique value is provided, the same percentage of transparency will
be applied to all the colors. If it is a vector of the same length as
\code{color}, a given transparency will be applied to each color.}

\item{name}{optional name to give to the resulting colors.}
}
\description{
This function easily creates transparent colors providing a color name and
the desired proportion of transparency
}
\examples{
plot(rnorm(1:100), rnorm(1:100), pch = 16, cex = 2, col = tCol("black", 70))
points(rnorm(1:100), rnorm(1:100), pch = 16, cex = 2, col = tCol("red", 40))
points(rnorm(1:100), rnorm(1:100), pch = 16, cex = 2, col = tCol("blue", 60))


}
