% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinData.R
\name{thinData}
\alias{thinData}
\title{Thin acoustic telemetry data}
\usage{
thinData(track.data = NULL, tag.id, rec.id, time.stamp, depth = NULL,
  time.int = "30min", depth.range = NULL, parallel = FALSE)
}
\arguments{
\item{track.data}{\code{data.frame} frame with acoustic telemetry data
containing the following columns: \code{tag.id}, \code{rec.id},
\code{time.stamp}, and \code{depth}. These arguments can be provided
separately.}

\item{tag.id}{vector with the ID codes of tagged individuals.}

\item{rec.id}{vector with the ID codes of the receivers for each detection.}

\item{time.stamp}{vector with the date and time of each detection in
\code{POSIXt} format.}

\item{depth}{vector with the depth values corresponding to each detection.}

\item{time.int}{time interval to which data will be thinned, as given to the
\code{units} argument of the \code{\link[lubridate]{round_date}} function
in the \code{lubridate} package.}

\item{depth.range}{vector with the minimum and maximum depth values,
respectively, between wich the sampling of depths will occur.}

\item{parallel}{option to apply the function in parallel (if \code{TRUE}).
This will be passed to the \code{.parallel} argument of the
\code{\link[plyr]{ldply}} function in the \code{plyr} package.}
}
\value{
Data frame with 4 columns: tag.id, time.stamps, rec.id, and depth.
}
\description{
This function thins acoustic telemetry data into specified time intervals.
For each interval, a receiver ID an a depth value are assigned depending on
their probability distribution in the original data.
}
