% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatChronPlot.R
\name{timeAxis}
\alias{timeAxis}
\title{Plot the Y axis of a spatial chronogram plot (times)}
\usage{
timeAxis(m)
}
\arguments{
\item{m}{matrix with the data to plot in the spatial chronogram plot.}
}
\value{
Plots the Y axis in a spatial chronogram plot.
}
\description{
Plots the time of the day in the Y axis of an spatial chronogram plot.
}
