% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracking}
\alias{tracking}
\title{Passive acoustic telemetry data of common dentex in the Medes Islands}
\format{Data frame with the following columns:
    \describe{
        \item{\code{tag.id}}{ID of the tagged individual}
        \item{\code{time.stamp}}{date and time of each detection, in
            \code{POSIXt} format}
        \item{\code{rec.id}}{ID of the acoustic receiver that registered
            each detection}
        \item{\code{depth}}{depth of the fish, provided by the acoustic
            transmitters in the coded signals}
    }}
\source{
Aspillaga, E., 2017. \emph{Movement ecology of coastal fishes in a marine
protected area: implications for management and conservation}. Doctoral
thesis. Universitat de Barcelona. 218 pp.
}
\usage{
tracking
}
\description{
Data frame with passive acoustic telemetry monitoring data of two common
dentex (\emph{Dentex dentex}) individuals (\code{'dentex18'} and
\code{'dentex43'}) in the Medes Islands marine reserve (Catalonia, NW
Mediterranean Sea). The tracking period took place from 01/Oct/2007 to
31/Dec/2007. Movements were tracked with a fixed array of 16 acoustic
\link{receivers}.
}
\keyword{datasets}
