% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeUD.R
\name{ud3dmesh}
\alias{ud3dmesh}
\title{Compute 3D contour mesh}
\usage{
ud3dmesh(rast.vol, levels = c(0.5, 0.95))
}
\arguments{
\item{rast.vol}{a \code{RasterStack} or \code{RasterBrick} object with
UD contour volumes.}

\item{levels}{vector with the probabilities of the contours to generate.}
}
\value{
A \code{list} with the 3D contour meshes.
}
\description{
This function takes a a \code{RasterStack} object with UD volumes
(generated by the \code{predictKde} function) and generates a 3D mesh to
be ploted.
}
