% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viewshed}
\alias{viewshed}
\title{Viewshed analysis output}
\format{A \code{RasterStack} object with one layer for each acoustic
    receiver.}
\usage{
viewshed
}
\description{
Output of the viewshed analysis performed for the acoustic receivers
(\code{\link{receivers}}) placed around the Medes Islands MPA (Catalonia, NW
Mediterranean Sea), using a local bathymetry raster
(\code{\link{bathymetry}}). The analysis was performed using the \code{GRASS}
software through the \code{rgrass7} package.
}
\keyword{datasets}
