% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap3d.R
\name{volOverlap}
\alias{volOverlap}
\title{Spatial overlap between two utilization distributions}
\usage{
volOverlap(ud1, ud2, level, symmetric = TRUE)
}
\arguments{
\item{ud1, ud2}{\code{RasterLayer}, \code{RasterStack} or \code{RasterBrick}
objects with the UD volumes to overlap. If it is a \code{RasterStack} or
\code{RasterBrick} object, the number and name of the layers must
coincide.}

\item{level}{UD volume probability contour to be used to calculate the
volume overlap.}

\item{symmetric}{logical. If \code{TRUE}, the overlapped index is calculated
referred to the total joint volume of the two UDs (volume(overlapped) /
volume(\code{ud1}) + volume(\code{ud2})). If \code{FALSE}, two overlap
indexes are calculated, the first one referred to the volume of
\code{ud1} (volume(overlapped) / volume(\code{ud1})), and the second one
referred to the volume of \code{ud2} (volume(overlapped) /
volume(\code{ud2})).}
}
\value{
A vector with one (if \code{symmetric == TRUE} or two
    (\code{symmetric == FALSE}) overlap indexes.
}
\description{
This function takes two \code{RasterLayer}, \code{RasterStack} or
\code{RasterBrick} objects with UD volumes and calculates the proportion
of the volume that is overlapped within a probability contour.
}
