% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumeUD.R
\name{volumeUD}
\alias{volumeUD}
\title{Calculate utilization distribution probability volumes}
\usage{
volumeUD(ud)
}
\arguments{
\item{ud}{a \code{RasterLayer} (2D), \code{RasterStack} or
\code{RasterBrick} (3D) object with UD values.}
}
\value{
a \code{RasterLayer} or a \code{RasterStack} object with the UD
    probability volumes.
}
\description{
This function calculates the utilization distribution probability volumes
from 2D or 3D UD values.
}
