% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelize.R
\name{voxelize}
\alias{voxelize}
\title{Voxelize list of trajectories}
\usage{
voxelize(synt.list, raster, depth.int, max.lag = 24, time.step = 2,
  parallel = FALSE)
}
\arguments{
\item{synt.list}{list with data frames containing the 3D trajectories to
process. Each element of the list must be a \code{data.frame} containing
the columns \code{x}, \code{y}, \code{depth}, and \code{time.stamp}
columns.}

\item{raster}{raster dataset (\code{RasterLayer} object) used to establish
the horizontal dimension and resolution of the 3D grid.}

\item{depth.int}{sequence of depth values indicating the breaks into which
the vertical axis is cut.}

\item{max.lag}{maximum time period, in hours, above which the segment
between two locations will not be included in the rasterization.}

\item{time.step}{time interval, in minutes, into which the trajectory will
be interpolated before the rasterization. This parameter is given to the
\link{interpPath} function.}

\item{parallel}{option to apply the function in parallel (if \code{TRUE}).
This will be passed to the \code{.parallel} argument of the
\code{\link[plyr]{ldply}} function in the \code{plyr} package.}
}
\value{
The function returns a \code{RasterStack} object, with one layer for
    each depth interval. The name of each layer corresponds to the mean depth
    of the interval.
}
\description{
This function takes a list of 3D trajectories and calculates the average
proportion of time that they expend in each voxel of the grid in which the
study area is divided.
}
